fix/folder structure for custom validation
fixes folder structure for custom validation error handler by add custom validation middlewares files to middleware folder and delete custom validation folder
