+++ auth/middleware/customLoginMiddleware.js
@@ -1,4 +1,4 @@
-import userMdl from "../../../schemas/userMdl.js";
+import userMdl from "../../schemas/userMdl.js";
 
 const emailExistingCheck = async (email) => {
   const countEmailExisting = await userMdl.user.countDocuments({ email });
@@ -49,20 +49,7 @@ const customLoginValidation = async (req, res, next) => {
 
   next();
 };
-const validate = (schema) => {
-  return async (req, res, next) => {
-    try {
-      await schema.validateAsync(req.body, { abortEarly: false }); // abortearly for stopping code to be exit when one error occur then show all errors in array in object
-      next();
-    } catch (error) {
-      const errorMessage = error.details.map((detail) => detail.message);
-      res.status(400).json({ error: errorMessage });
-      console.log(errorMessage);
-    }
-  };
-};
 
 export default {
   customLoginValidation,
-  validate,
 };
