+++ auth/middleware/customRegisterMiddleware.js
@@ -1,4 +1,4 @@
-import userMdl from "../../../schemas/userMdl.js";
+import userMdl from "../../schemas/userMdl.js";
 
 const emailExistingCheck = async (email) => {
   const countEmailExisting = await userMdl.user.countDocuments({ email });
