+++ auth/register/registerService.js
@@ -1,26 +1,26 @@
 import userMdl from "../../schemas/userMdl.js";
 import utils from "../utils/utils.js";
 
-// const emailExistingCheck = async (email) => {
-//   const countEmailExisting = await userMdl.user.countDocuments({ email });
+const emailExistingCheck = async (email) => {
+  const countEmailExisting = await userMdl.user.countDocuments({ email });
 
-//   if (countEmailExisting > 0) {
-//     return true;
-//   }
-//   return false;
-// };
+  if (countEmailExisting > 0) {
+    return true;
+  }
+  return false;
+};
 
 const registerUser = async (req, res, next) => {
   try {
     const { name, email, password } = req.body;
 
-    // const emailExistCheck = await emailExistingCheck(email);
+    const emailExistCheck = await emailExistingCheck(email);
 
-    // if (emailExistCheck) {
-    //   const error = new Error("USER ALREADY EXIST");
-    //   error.status = 400;
-    //   throw error;
-    // }
+    if (emailExistCheck) {
+      const error = new Error("USER ALREADY EXIST");
+      error.status = 400;
+      throw error;
+    }
 
     const userData = await userMdl.user({
       name: name,
